% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fwi.R
\name{plot_fwi}
\alias{plot_fwi}
\title{Plot bushfire climate driver over time}
\usage{
plot_fwi(data, var = "temp_c")
}
\arguments{
\item{data}{A data frame, typically \code{bushfire_real}.}

\item{var}{Variable to plot: one of "temp_c", "wind_ms", "dryness_index".}
}
\value{
A ggplot2 line plot.
}
\description{
Plot temporal trends of temperature, wind, or dryness
from the \code{bushfire_real} dataset.
}
\examples{
data(bushfire_real)
plot_fwi(bushfire_real, var = "temp_c")
}
